==> notebook_examples/0.dataset.ipynb <==
import pyarrow.parquet as pq
import pyarrow as pa
import ujson

# 1. Process Wiki data

origin_wiki_file = './data/wiki.simple.txt'

liness = []
with open(origin_wiki_file, 'r', encoding='utf-8') as f:
    lines = f.readlines()

lines[0:5]

# Delete empty lines
new_lines = []
for line in lines:
    if line.strip() != '':
        new_lines.append(line)

print(len(lines), len(new_lines))

# with open('./data/wiki.simple_new.txt', 'w', encoding='utf-8') as f:
#     f.writelines(new_lines)

tb = pa.Table.from_arrays([new_lines], names=['text'])
# compression='GZIP'
pq.write_table(table=tb, where='./data/wiki.simple.parquet', row_group_size=50000, data_page_size=50000, )

# 2. Process training data

my_data = []
max_len = 384

with open('./data/bell_open_source/train_3.5M_CN.json', 'r', encoding='utf-8') as f:
    for line in f:
        item = ujson.loads(line)

        if len(item['conversations']) != 2: continue

        conversation = item['conversations']
        txt = ''
        if conversation[0]['from'] =='human':
            txt = f"{conversation[0]['value']}\n{conversation[1]['value']}"
        else:
            txt = f"{conversation[1]['value']}\n{conversation[0]['value']}"

        if len(txt) >= max_len: continue
        my_data.append(txt)


for file in ['./data/bell_open_source/train_2M_CN.json', './data/bell_open_source/Belle_open_source_1M.json']:
    with open(file, 'r', encoding='utf-8') as f:
        for line in f:
            item = ujson.loads(line)

            if item['input'].strip() != '':
                txt = f"{item['instruction']}\n{item['input']}\n{item['output']}"
            else:
                txt = f"{item['instruction']}\n{item['output']}"
            
            if len(txt) == 0 or len(txt) >= max_len: continue
            my_data.append(
                    txt
            )

print(len(my_data), my_data[0:5])


tb = pa.Table.from_arrays([my_data], names=['text'])
# compression='GZIP'
pq.write_table(table=tb, where='./data/bell_pretrain_3M.parquet', row_group_size=20480, data_page_size=20480, )

# 3. Process sft data

lines = []
with open('./data/sft_0.8M_CN.json', 'r', encoding='utf-8') as f:
    for line in f:
        item = ujson.loads(line)

        txt = f"{item['instruction']}{item['output']}"
        
        if len(txt) == 0 or len(txt) >= 320: continue
        lines.append(
                item
        )

print(len(lines))

tb = pa.Table.from_pylist(lines)
# compression='GZIP'
pq.write_table(table=tb, where='./data/sft_train_data.parquet', row_group_size=20480, data_page_size=20480, )
