==> notebook_examples/5.chat_demo.ipynb <==
from transformers import PreTrainedTokenizerFast, GenerationConfig , PhiForCausalLM, pipeline
import torch

# 模型保存路径
这里加载sft后的模型和dpo后的模型，看看两者输出的区别

tokenizer_dir = './model_save/tokenizer/'
dpo_model_save_dir = './model_save/dpo/'
sft_model_save_dir = './model_save/sft/'
max_seq_len = 320

## 加载模型

device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
print(device)

tokenizer = PreTrainedTokenizerFast.from_pretrained(tokenizer_dir)
dpo_model = PhiForCausalLM.from_pretrained(dpo_model_save_dir).to(device)

sft_mdodel = PhiForCausalLM.from_pretrained(sft_model_save_dir).to(device)

model_size = sum(t.numel() for t in dpo_model.parameters())
print(f"Phi-2 size: {model_size / 1000**2:.1f}M parameters")

## 定义两个text-generation的pipeline

dpo_pipe = pipeline(
    "text-generation", 
    model=dpo_model, 
    device=device,
    tokenizer=tokenizer,
)

sft_pipe = pipeline(
    "text-generation", 
    model=sft_mdodel, 
    device=device,
    tokenizer=tokenizer,
)

## 模型输出

txt = '感冒了要怎么办？'
prompt = f"##提问:\n{txt}\n##回答:\n"
sft_outputs = sft_pipe(prompt, num_return_sequences=1, max_new_tokens=256, pad_token_id=tokenizer.eos_token_id)
dpo_outputs = dpo_pipe(prompt, num_return_sequences=1, max_new_tokens=256, pad_token_id=tokenizer.eos_token_id)


print(f"model after sft output:\n\n{sft_outputs[0]['generated_text']}")
print('\n==================\n')
print(f"model after dpo output:\n\n{dpo_outputs[0]['generated_text']}")

## 不用pipeline组件，使用greedy search方法手动生成

# greedy search
gen_conf = GenerationConfig(
    num_beams=1,
    do_sample=False,
    max_length=320,
    max_new_tokens=256,
    no_repeat_ngram_size=4,
    eos_token_id=tokenizer.eos_token_id,
    pad_token_id=tokenizer.pad_token_id,
)

tokend = tokenizer.encode_plus(text=prompt)
input_ids, attention_mask = torch.LongTensor([tokend.input_ids]).to(device), torch.LongTensor([tokend.attention_mask]).to(device)

outputs = dpo_model.generate(
    inputs=input_ids,
    attention_mask=attention_mask,
    generation_config=gen_conf,
)

outs = tokenizer.decode(outputs[0].cpu().numpy(), clean_up_tokenization_spaces=True, skip_special_tokens=True,)
print(outs)
